refactor(mentor): 댓글 스크롤 hook 생성
