+++ components/ArticleDetail/EditComment.js
@@ -1,6 +1,5 @@
 import Button from '@/utils/Button';
 import styles from '@/styles/Comment.module.css';
-
 import { useState } from 'react';
 import { editCommentApi } from '@/utils/api/commentApi.js';
